Converted find-games to a vue-formatted file
